+++ test/gpui/schema_test.exs
@@ -6,6 +6,9 @@ defmodule GPUI.SchemaTest do
              :div,
              :button,
              :ui_button,
+             :ui_popover,
+             :ui_popover_trigger,
+             :ui_popover_content,
              :ui_checkbox,
              :ui_input,
              :ui_select,
@@ -30,6 +33,7 @@ defmodule GPUI.SchemaTest do
     assert :"phx-search" in GPUI.Schema.events()
     assert :"phx-release" in GPUI.Schema.events()
     assert :ui_select in GPUI.Schema.identified_tags()
+    assert :ui_popover in GPUI.Schema.identified_tags()
     assert :raster in GPUI.Schema.resources()
     assert :border_radius in GPUI.Schema.styles()
     assert :font_weight in GPUI.Schema.styles()